requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
supabase==2.9.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# PDF Generation
reportlab==4.2.5
Pillow==11.0.0

# Image Storage
cloudinary==1.41.0

# Environment & Configuration
python-dotenv==1.0.1

# Utilities
python-multipart==0.0.12
aiofiles==24.1.0
email-validator==2.2.0

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Optional: Enhanced logging
structlog==24.4.0
pytest
pytest-asyncio
httpx
